Remove mobileClient
