+++ UML/Gerer_Produits/diagramme_sequence_gererProduits_details.xml
@@ -1,0 +1,1 @@
+<mxfile host="www.draw.io" modified="2019-10-28T13:50:16.866Z" agent="Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:70.0) Gecko/20100101 Firefox/70.0" etag="7Lk36XycFwlpEzNx0hWN" version="12.1.7" pages="1"><diagram id="-etBZFAO_1bb2ClVudc-" name="Page-1">7Vtdc5s4FP01nu4+JIMQYPNoJ03b2WY303Rmt087Csi2Ghl5sEic/fUrgTAIyYa40Dht8+CgC0hw7zn3S/YIXqy271K0Xl6zGNOR68TbEbwcueLPC8U/KXkqJOMQFIJFSuJCVBPckv+wEjpKmpEYb7QLOWOUk7UujFiS4IhrMpSm7FG/bM6ovuoaLbAhuI0QNaV/k5gvC+nEHVfy95gsluXKIFAvvELlxepNNksUs8eaCL4dwYuUMV4crbYXmErllXop7rvac3b3YClOeJcbcHr14evNzPe9zfWfsy9/Ifzx6kzN8oBopl5YPSx/KjUgnnstD7MV/UjmmJJEjGZiXU4iskb54uLcNOIslSdwSlaYY3F8SdX1N5Vs9rgkHN+uUSTnfBSgEbIlX1ExAuJQ2JEjcUu6G1OK1htylz+OIyQpjrJ0Qx7wJ7wp4CKlD1g+D6JTShaJkHEmJ96IdUiy+CwHlzAQEoruMJ2h6H6RsiyJLxiVTy3eE87zP3EJy7h86IsdouT8ailfHpuaL9UoHgJvayJliXeYiddPn8QlW+0GRQroesX4sYJYaYmlhi4QKGgrWC92M1eWFwfK+M8AgtsOhJqV1owkPH8EfzbyLxtmZylfsgVLEK0b3tTaQUR2VuVEU+XYMTRZupK6JsPyut4VCQ1FTm8+CMGMUHzN+lZqF3LmkBfSuaDWFVoRKlX1HtMHLBmjTijPC1w1rhHDyf+EHCluRcKCxeo8ZffYdumcUGrjlp287aTvgXKuH+ik802ohJ6FdE44FFa8UyEdfK4yPV/T5aQj67yhNOm/Wk36uiaB43dS5WCaDAxNPowu4GgapmROhCZcRyhUJmH3ODF0jGORPKlhpca3lbQe6kt3QvGcV7H6Yz66lH4IJ/FUpnBieEdZdC/9hwzaOFZeQZgiffpHTZcPvuz1F4esxVG6wIeuUzaRr3fQpimmiAvvpqehFhOpW28kECssBI7OKl+fYMOyNMLqnsrMQkfoqXaZQvfeVSa+bZUKM8V8FYJ279cJVOs/z959viPXX7f/3rJH/9Mftyw+mxigQlToe1oBSZ4Tmf6h7PMqRSvcnkGWxAFjKE+UnHElLGK0WeboAbvoVWb1cIcag7tdgbSX42cu1GOPH5oO0w8Dk+buYLEHmBmfUMoURVFB9zjnen7sSNIj+UE2XJJfVmRCKSmLM1IgbZ+nbabld4xztrJze2+Mb6VdTYtlJPpGJjYSdA92o6IxD3SCwxMVrseY6Aji2a1spqMXlGC1WN1oAvJcJ1IqUzRU5WBm1EQZZyqNszj0DiWZyIB6ye3OgO454Xhs8ssSRd3B2GWmdjO0KbgjP0RtWdAMH6TPr5Jg0JLAG4c6bCwlAXCAiRsAnMFqAhM54BxIV5yHy7SATZStVVZWpGQv4Z0PAv8FfPYZCPReABgf6bU99/A8Aztt72SqmZ0b604pr6G6bpVhH+0Ya+ZpNjjZ2gx+ryTLLMyxPwp6vk4ALzCjYDgxte+V/cf+24pjQ/8nl/m7PfWX9JTRtxTyvgctsWSwFN81yy5wDqtQgpMHRvCw4YOthY0b6henVW4x+aa0XyneOXdgCHsJISJbOder4yBoUKNrDHFbZ9oTRZ5bzgsXu2cla0lv3q8/FZvPN3iYsGa2lgxfkLAf1yOMXUvRb/MIk8FyS9Mbg3PP4hCyRH7iNMWZTC9/K8wiy72UbIrE8/efzTNA3ZhNI3X2Cy3z9OYV7Oucmk+whSi3hkgu1Jr8ilE12/iu3uvwgyORCEHLRD1B0Qvs67wcFK2VgtkpO4Sr44uulsT/MI4g0Gk9Cb/fdpFVa2a34vVuFzXIbCV8fatoP4x63xmCQK/sJh1bG8/lKvSgdZ1Bd4fMdodr7kSgbCsdfyHiiFAZAeI3WVIFgm+NA3qrqx0N32Ujwm6PIzYiWgDUX0/LamMz95c2BgebmnmQJ8mc/RymhuWW3O6rAM0mc1dj+27bTANb26wzTjCW+mXXS2kp7BhLB+tVhiZH7O0arHtCOcr640bXdPgY1gxRmLnnza/QNC3UuTZrn6q38mzvUieWFpcMOP2uTefeekvXRo+4wLVsCvrh2PQLvewJ2m1g7mLk8Dnpvs2puIeGOWFzz//Ixo050UCdm91Cp+YYXAso7a0bS7zqNZt7XZA0s7Ojv11kZIwdv170/BbOnoVODZNmKbkLVjusiSmE0cmcFIh8uYDlDhKwwrFl49ENLfEKDBavDuz01MyQyB9tZT+a/gGw7fPYDNDHPo9d/2b95Xba56mMk7zBW9VsX6PNT5c0gIZJx43fXXVOGtom6i1p2LPQd3PQYlj9lrG4vPpFKHz7Pw==</diagram></mxfile>
