+++ UML/Gerer_Produits/diagramme_sequence_gererProduits_liste.xml
@@ -1,1 +1,1 @@
-<mxfile host="www.draw.io" modified="2019-10-17T16:00:23.302Z" agent="Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:69.0) Gecko/20100101 Firefox/69.0" etag="CsPpecHN4qL8CnQm-3dF" version="12.1.3" type="device" pages="1"><diagram id="-etBZFAO_1bb2ClVudc-" name="Page-1">7Vnfc+I2EP5rmGkfkvFv4BFI6N1M0maam2nvUbEXUCMsRpZJ0r++K1vGFrKBS/ANbY8HsNbySvr2208rM/Bn69dfBNms7nkCbOA5yevAvxl4+AnG+KMsb6VlOHZLw1LQpDQ1DI/0b9BGR1tzmkBmdJScM0k3pjHmaQqxNGxECP5idltwZo66IUuwDI8xYbb1D5rIVWkdecPa/gnoclWN7EZ6wWtSddYryVYk4S8Nk3878GeCc1lerV9nwBR4FS7lc/OOu7uJCUjlKQ+AmH/+62EahkF2/+v0628E7uZX2suWsFwvWE9WvlUI4Lw36jJfszu6AEZTbE1xXEljuiHF4HhvEksu1A0QdA0S8PqG6f4PtW36sqISHjckVj5fkDRoW8k1w5aLlxhHSfARsWszRjYZfSqm46BFQJyLjG7hd8hKuijrFtR8CJswukzRJrlynOE4NF1+UY0bP0ILI0/ApiR+Xgqep8mMMzVrXKe/KD7YhedSTXq2Y5Tyr4cK1bXGDEeE185guLsQY24Ax+WLN+xiPqCTwg80R15qilWRWDXYFQ51P6JZvdw5rgOPFzr238AD7zgPGkHacJrKYgrhdBDe7EWdC7niS54S1oz7SaB1c7QTyZGB5NAGslKSJpBBb0D6FpCTh89omFIG9/zcoJ6SmwXj0brAzJqTNWUKqk/AtqASRt/Qwut6ut3IC6f4oJ3o1IoxXOXoUvBnaOu6oIy1pVZ77h7P+TNknBdGZs6FNlXGw5acC/qiSnDBOecfxDIIDShHJyZdb0iG/1okQxNJ1wlPg7IvJCMLye1g5g8mY0EXFJHwHARUlWDPkFoYQ4Klk27WMN7W1uZGX6kJg4Wsd+q7onWjZAjSZKIKOGw+MR4/K/lQWzYkWhQQd/H2p3ZXNL4acqGm845Y4eqIWMKhfmF7TAUwIlHczCK0JUT60QdFxJoLkWNmVWg6yHguYtDP1GFGjMhbo5tmd+coo7BtlJozpb+aQbv1vZ9Url1dYMgmJI5LbiUFsYprRzGMqC+aScU0Vfxj5ARPclouqyut9yvAJy4lX7cT6QSCdOwmjcSsZO+DYd+rBQP/tLhbfnwnOuyopLXl6GxRtkufGaOgB2sGDTVUmmIgVDlA6v3elmiSS65Lhhb1OKH6x+32LHXElWumqT/2LckOWyTb60uyXbuOmJKszB31hceYMs3gYPr8KD97LT+DoXlQuYDy0+aNe+0qIfYnStVK0sT5RhcA5e5/Edp8IA++h2JfuZEZTHf4Ts0OvMN+epbs4JIL5654VtAFe9Cddgjp7w3K0MKSMFnmki6X1T2qRLn7DdtckDUcf0tWLdId+upGtT6Vu9OEZKuiRnZ3Glm9ufTOdJQ2K5aw5dASBr6NvTvuDfxRi5b5tZZBuuUU+tUvvsEY78GPt/XWNvpQ1amBd66dXbnxQQ3DzfLaPAlEUfA+EfOOeuqQsW89umCOd4zUenyxnzdnxReLDPrRVfsYbWlByv+7ijD0bDFuVYRRb8WNrcbuddAiCHmqvkEIyFV9EziqAvqpDI46cwialfXPz/83ffDNkO6H6mR1OOLnbNrQPs6lKUPbRuU1eCkR1vTHTtWITeiZB+4weicTffeIozNRMYjax/luVMRm/bdy2b3+c96//Qc=</diagram></mxfile>
+<mxfile host="www.draw.io" modified="2019-10-28T13:30:13.669Z" agent="Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:70.0) Gecko/20100101 Firefox/70.0" etag="EkGwm4G-gd9b86EYkHGa" version="12.1.7" pages="1"><diagram id="-etBZFAO_1bb2ClVudc-" name="Page-1">7VnbbuM2EP0aA+1DAt1tP/oS7y6QtEGzQLuPjDS22dCiQVF20q/vUKIs0ZQvSaxu2q4fbHFEDckzZw6Hcs+frJ4/CbJe3vEEWM9zkueeP+15+AmG+KMsL6WlP3RLw0LQpDQ1DA/0L9BGR1tzmkBmdJScM0nXpjHmaQqxNGxECL41u805M0ddkwVYhoeYMNv6O03ksrQOvH5t/wx0saxGdiO94BWpOuuVZEuS8G3D5N/0/IngXJZXq+cJMAVehUv53OzA3d3EBKTynAdAzL78eT8OwyC7+2X87VcCt7Mr7WVDWK4XrCcrXyoEcN5rdZmv2C2dA6MptsY4rqQxXZNicLw3iiUX6gYIugIJeD1luv99bRtvl1TCw5rEyucWSYO2pVwxbLl4iXGUBB8RuzZjZJ3Rx2I6DloExLnI6AZ+g6yki7JuQM2HsBGjixRtkivHGY5D08VX1Zj6EVoYeQQ2JvHTQvA8TSacqVnjOv158cEuPJdq0pMdo5R/PVSorm3kKxhxEvDcMOlIfAKOyxcv2OXZeEAnhR9ojmxrilWRWDbYFfZ1P6JZvdg5rgOPFzr2r+CBd5oHjSCtOU1lMYVw3Aune1HnQi75gqeENeNug3aUkGcjOTCQ7NtAVkrSBDLoDEjfAnJ0/wUNY8rgjl8a1HNys2A8WueYWTOyokxB9RnYBlTC6BtaeF1Ptxt54RQftBOdWjFGsBxdCv4EbV3nlLG21GrP3dM5f4GM88LIzLnQpsqw35JzQVdUCT5KzvmvxTIIDSgHZyZdZ0iG/1okQxNJ1wnPg7IrJCMLyU1v4vdGQ0HnFJHwHARUlWBPkFoYQ4Klk27WMN7U1uZGX6kJg7msd+rbojVVMgRpMlIFHDYfGY+flHyoLRsSLQoYCvHyh3ZXNL4dlItj0ZJELOBYPx0TtbyjMRXAiERxM4vQlhDpR+8VEWsuRI6ZVaHpIOO5iEE/U4cZMSIvjW6a3QdHGYRto9ScKf3VDNqt7+2kcu3qAkM2InFccispiFVcO4phRH3RTCqmqeIfQyl4ktNyWYfSer8CfORS8lU7kQ7uJydj3EjMSvbeGfa9WjDwz4u75cd3ouOOSp5bji4WZbv0mTAKerBm0FAbpSkGQpUDpN7vbYkmueS6ZGhRjzOqf9xuL1JHXLlmmvpD35LssEWyva4k27XriDHJytxRX3iMKdMMjqbPj/Kz0/Iz6JsHlQ9Qftq8ca9dJcT+SKlaSZo4X+sCoNz9v4c2H6X9d1DsKzcyg+n236jZgXfcT8eSHXyYwnknYucnVLAH3XmHkO7eoPQtLAmTZS7pclndo0qUD79hmwmygtNvyapFun1f3ajWp3J3nJBsWdTI7k4jqzeX3oWO0mbFErYcWsLAt7F3h52BP2jRMr/WMkg3nEK3+sXXGOM9+PG23toG76o6NfDOtbMrN96pYbhZXpsngSgK3iZi3klPB2TstUcXzPEDI7UeX+znzVnx+TyDbnTVPkZbWpDy/64i9D1bjFsVYdBZcWOrsXsdtAhCnqpvEAJyVd/8VIZFnTYEzcrK5+f/mzL4ZjD3g3S2LpzwczFVaB/no2lC2xblNRgpEdb0xx7ViE3omUftMHojE333hKMLUTGI2sf5x6iIzfoP5bJ7/be8f/M3</diagram></mxfile>
