+++ src/App/EventListener/AuthenticationSuccessListener.php
@@ -20,10 +20,12 @@ class AuthenticationSuccessListener
             return;
         }
 
+        date_default_timezone_set("Europe/Paris");
         $data['data'] = array(
             'Message' => "Bienvenue " . $user->getName() . ", content de vous retrouver !",
             'name' => $user->getName(),
             'roles' => $user->getRoles(),
+            'expiresAt' => date("d/m/Y H:i:s", time() + 3600),
         );
 
         $event->setData($data);
